Update fake.login.js
After a couple time trying to follow the tutorial, and getting this error over and over: "Runtime error encountered: You probably have multiple tabs open to the same origin." i solved the issue, and it had to do with the way i was writting this file. It has to be await Promise.all, and the page.close() is mandatory in the end. Btw, greate tutorial and work!